# MF_API_document.yml
openapi: 3.1.0
...
info:
  title: Money Forward Invoice API
  version: 3.5.0
  summary: Money Forward Invoice API
  description: |-
    ### マネーフォワード クラウド請求書APIについて
    https://biz.moneyforward.com/support/invoice/guide/api-guide/a03.html
    ### マネーフォワード クラウド請求書API v3 スタートアップガイド
    https://biz.moneyforward.com/support/invoice/guide/api-guide/a04.html
servers:
  - url: /api/v3
paths:
...
  '/partners/{partner_id}':
    parameters:
      - schema:
          type: string
        name: partner_id
        in: path
        required: true
    get:
      summary: Get a partner
      description: 取引先の取得
      tags:
        - Partner
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
              examples:
                with-departments:
                  $ref: '#/components/examples/PartnerWithDepartment'
                without-departments:
                  $ref: '#/components/examples/PartnerWithoutDepartment'
        '404':
          description: Not Found
      operationId: get-partners-id
    put:
      summary: Update a partner
      description: 取引先の更新
      operationId: put-partners-id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
              examples:
                Example:
                  $ref: '#/components/examples/PartnerWithDepartment'
        '400':
          description: Bad Request
      requestBody:
        $ref: '#/components/requestBodies/PartnerUpdateRequest'
      tags:
        - Partner
      security:
        - AccessToken:
            - mfc/invoice/data.write
    delete:
      summary: Delete a partner
      description: 取引先の削除
      operationId: delete-partners-id
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
      tags:
        - Partner
      security:
        - AccessToken:
            - mfc/invoice/data.write
...
components:
  schemas:
    Partner:
      title: Partner
      x-stoplight:
        id: mnavdatyrxciq
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        name_kana:
          type: string
        name_suffix:
          type: string
        memo:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        departments:
          type: array
          items:
            $ref: '#/components/schemas/Department'
      required:
        - id
        - name
        - created_at
        - updated_at
        - departments
...
    Department:
      title: Department
      x-stoplight:
        id: bq8upk7nj10bc
      type: object
      properties:
        id:
          type: string
        zip:
          type: string
        tel:
          type: string
        prefecture:
          type: string
        address1:
          type: string
        address2:
          type: string
        person_name:
          type: string
        person_title:
          type: string
        person_dept:
          type: string
        email:
          type: string
        cc_emails:
          type: string
        peppol_id:
          type: string
        office_member_id:
          type: string
        office_member_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
...
  requestBodies:
...
    PartnerUpdateRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                maxLength: 30
              name:
                type: string
                maxLength: 100
              name_kana:
                type: string
                maxLength: 350
              name_suffix:
                type: string
              memo:
                type: string
                maxLength: 500
      description: Request body for updating a Partner
...
  securitySchemes:
    AccessToken:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: 'https://api.biz.moneyforward.com/authorize'
          tokenUrl: 'https://api.biz.moneyforward.com/token'
          scopes:
            mfc/invoice/data.write: Grant read and write access to all your office's data
            mfc/invoice/data.read: Grant read-only access to all your office's data
          refreshUrl: 'https://api.biz.moneyforward.com/token'
...
  examples:
...
    PartnerWithoutDepartment:
      value:
        id: 95PHKI9_FeSw3coTj673Cg
        code: p41uz1dyvw3cj71qrkja
        name: Ryu p41uz1dyvw3cj71qrkja
        name_kana: p41uz1dyvw3cj71qrkja
        name_suffix: 御中
        memo: p41uz1dyvw3cj71qrkja
        created_at: '2023-03-20 13:39:28 +0900'
        updated_at: '2023-03-20 13:39:28 +0900'
        departments: []
    PartnerWithDepartment:
      value:
        id: 95PHKI9_FeSw3coTj673Cg
        code: p41uz1dyvw3cj71qrkja
        name: Ryu p41uz1dyvw3cj71qrkja
        name_kana: p41uz1dyvw3cj71qrkja
        name_suffix: 御中
        memo: p41uz1dyvw3cj71qrkja
        created_at: '2023-03-20 13:39:28 +0900'
        updated_at: '2023-03-20 13:39:28 +0900'
        departments:
          - id: qwc4iT7ZrywxipJCOqtZQg
            zip: 123-4567
            tel: '1234567'
            prefecture: 山形県
            address1: hb3m8kaxz9eex1czmpn2
            address2: hb3m8kaxz9eex1czmpn2
            person_name: hb3m8kaxz9eex1czmpn2
            person_title: hb3m8kaxz9eex1czmpn2
            person_dept: hb3m8kaxz9eex1czmpn2
            email: hb3m8kaxz9eex1czmpn2@moneyforward.com
            cc_emails: hb3m8kaxz9eex1czmpn2@moneyforward.com
            peppol_id: '0088:0000000000001'
            office_member_name: hb3m8kaxz9eex1czmpn2
            office_member_id: '-UNhHGbLKnWH5xlrFhj2ow'
            created_at: '2023-03-20 13:44:52 +0900'
            updated_at: '2023-03-20 13:44:52 +0900'
...
security:
  - AccessToken:
      - mfc/invoice/data.write
      - mfc/invoice/data.read